+++ app/src/pages/Challenge/ChallengeCompletion/Confetti.tsx
@@ -1,0 +1,29 @@
+
+// Confetti animation component
+const Confetti = () => {
+  const confettiPieces = Array.from({ length: 50 }, (_, i) => ({
+    id: i,
+    left: Math.random() * 100,
+    animationDelay: Math.random() * 3,
+    color: ['bg-yellow-400', 'bg-blue-400', 'bg-green-400', 'bg-red-400', 'bg-purple-400'][Math.floor(Math.random() * 5)]
+  }));
+
+  return (
+    <div className="fixed inset-0 pointer-events-none overflow-hidden">
+      {confettiPieces.map((piece) => (
+        <div
+          key={piece.id}
+          className={`absolute w-2 h-2 ${piece.color} animate-bounce`}
+          style={{
+            left: `${piece.left}%`,
+            animationDelay: `${piece.animationDelay}s`,
+            animationDuration: '3s',
+            top: '-10px'
+          }}
+        />
+      ))}
+    </div>
+  );
+};
+
+export default Confetti;
