+++ app/src/pages/Challenge/ChallengeDetails/ChallengeDetail.tsx
