--- app/src/pages/LogIn.tsx
@@ -1,13 +1,0 @@
-import LoginForm from "@/components/LoginForm";
-
-function LogInPage() {
-
-  return (
-    <div className="min-h-screen flex-center p-6">
-    {/* <div className="min-h-screen bg-gradient-to-br from-orange-50 via-yellow-50 to-pink-50 flex items-center justify-center p-6"> */}
-      <LoginForm />
-    </div>
-  );
-}
-
-export default LogInPage;
