+++ app/src/pages/LogIn/LogIn.tsx
@@ -1,0 +1,12 @@
+import LoginForm from "@/pages/LogIn/LoginForm";
+
+function LogInPage() {
+
+  return (
+    <div className="min-h-screen flex-center p-6">
+      <LoginForm />
+    </div>
+  );
+}
+
+export default LogInPage;
